refactor: improve person tab le jsx structure and rework PersonItem compo nent
