+++ src/pages/PeoplePage.tsx
@@ -1,10 +1,13 @@
-import { People } from '../components/People/People';
+import { PeopleTable } from '../components/People/PeopleTable';
 
 export const PeoplePage = () => {
   return (
     <div className="container">
       <h1 className="title">People Page</h1>
-      <People />
+
+      <div className="block">
+        <PeopleTable />
+      </div>
     </div>
   );
 };
